Extract extractionToRecipe into shared module
Deduplicate the extractionToRecipe helper that was defined independently in
evaluate-extraction.ts, ExtractionListView.tsx, and StatsView.tsx. The shared
version in src/lib/extraction-to-recipe.ts accepts an ExtractionInput interface
compatible with both the Zod schema type and the viz interface type.
